Add zustand, Seperate screens for base image picker and palette generator.WIP:remove original index.tsx
