+++ app/(tabs)/result.tsx
@@ -1,0 +1,1 @@
+import { useRouter } from 'expo-router';
